Youtube/requirements.txt
google_api_python_client==2.86.0
isodate==0.6.1
mysql_connector_repackaged==0.3.1
pandas==1.5.3
pymongo==4.3.3
python_dateutil==2.8.2
streamlit==1.22.0
